refactor: Rename api to controller
